Simplified resource management.
